[CYTHON] Fix potential flaky issue in dispatcher
This PR fixes a flaky bug in dispatcher.

The C++ dispatcher dispatches the argument passing by TYPE(obj) pointer which
is non-owning. This means that there is a following edge case:
- type A is registered through dispatcher
- type A get garbage collected (because it is a local type)
- type B is created and use the same memory address as type A

Then when we pass in type B, it will mistakenly use dispatch function for type A

To prevent this, we keep alive the types that are registered through dispatcher
by appending them to _DISPATCH_TYPE_KEEP_ALIVE

NOTE that total number of types that are registered through dispatcher is expected
to be limited in practice so we can afford to keep them alive