[ABI] Further clarify Function ABI in cpp
This PR elevates the cpp_call part of FunctionObj to the c api.
We also allows the cpp_call to be nullptr. This change would
make the function ABI stablized.

This change is backward compatible with previous function ABI.
Previous functions can be passed to this ABI. But future functions
may contain cpp_call as NULL and cannot be passed to older dll.