[FEAT] Introduce TensorView for non-owning view
When exposing an FFI function, sometimes it is not always possible
to pass in an owning reference of a Tensor, in such case, it is better
to interface with the Tensor using non-owning view.

This PR adds a ffi::TensorView class to represent in-memory non-owning
DLTensor view. It can be converted from DLTensorPtr* in the ffi argument
but cannot be promoted to an owning Tensor.

We recommend kernel ops to make use of TensorView when possible so
the exposed ops can support broad range of inputs including non-owning ones.