[DLL] Explicit link libtvm_ffi_testing from cython
This PR changes the linking of libtvm_ffi_testing to explicit linking.
This helps to simplify the unloading order, to ensure cython always
get unloaded before libtvm_ffi_testing in case we need to de-allocate
global objects there. While it introduces one testing dep,
the behavior is similar as before and on python such extra dep
is ok for testing.