[ERROR] Make Error conform more to std
This PR changes Error.what() to only return message.
This helps to conform to the standard that what should not throw.
Added a error.FullMessage() function to get the fullmessage with trace and kind.

This also helps us to remove TLS in the header side which can makes the
compiled c++ object more friendly wrt to LLVM JIT.